Update randomizer.py
Updated for overflow fix when purchasing boards. Was checking with subtractions if the value hits zero previously. Now fixes it so it will only check if the value is greater than zero, it will continue subtracting. Temporary fix. You will need 500 coins rounded up to make the purchase or the coins counter will go into the negatives. Working for a fix for this...